requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
flask==2.3.3
pandas==2.1.1
plotly==5.18.0
dash==2.14.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
